+++ firecracker-pilot/firecracker-service/Cargo.toml
@@ -7,3 +7,5 @@ edition = "2021"
 
 [dependencies]
 daemonize = "0.5.0"
+env_logger = "0.10.0"
+log = "0.4"
